requirements.txt
streamlit
numpy
scikit-learn
joblib
streamlit-drawable-canvas
pillow
opencv-python-headless
gdown

